backend/requirements.txt
# F1 Race Strategy Explorer — Backend dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
fastf1>=3.3.0
pandas>=2.1.0
numpy>=1.26.0
pytest>=8.0.0
